Moved somme files and fixed imports
