+++ nextjs-blog/pages/profile/info.js
@@ -1,13 +1,13 @@
 import Head from 'next/head'
-import styles from '../styles/Home.module.css';
-import clientPromise from "../lib/mongodb";
+import styles from '../../styles/Home.module.css';
+import clientPromise from "../../lib/mongodb";
+// import Profile from '.';
+import Profile from '../profile/index';
 import { withPageAuthRequired, getSession } from '@auth0/nextjs-auth0';
-import Profile from './profile';
 import React, { useState } from "react";
 import { useTranslation } from "react-i18next";
-import "../src/Translation/i18n";
+import "../../src/Translation/i18n";
 import i18n from "i18next";
-
 
 export default function UserInfo({ user, data }) {
   const [language, setLanguage] = useState('en');
@@ -22,9 +22,9 @@ export default function UserInfo({ user, data }) {
       i18n.changeLanguage(e.target.value);
     }
   }
-    return (
-      user && (
-<div className={styles.container}>
+  return (
+    user && (
+      <div className={styles.container}>
         <Head>
           <title>{t("TitleProfile")}</title>
           <link rel="icon" href="/favicon.ico" />
@@ -57,41 +57,40 @@ export default function UserInfo({ user, data }) {
         </header>
         <main>
           <h1 className={styles.title}>
-          {t("AccountCre")}
+            {t("TitleProfile")}
           </h1>
-          <br/>
+          <br />
           <div className="container">
-                <div className="card mt-5">
-                    <form className="card-body" action="/api/forms/newClientForm" method="POST">
-                    <input type="hidden" className="form-control" value={data?.email} id="email" name="email"/>
-                        <div className="form-group mb-3">
-                        <label className="mb-2"><strong>{t("FirstName")}</strong></label>
-                            <input type="text" className="form-control" id="first_name" name="first_name" required/>
-                        </div>
-
-                        <div className="form-group mb-3">
-                        <label className="mb-2"><strong>{t("LastName")}</strong></label>
-                            <input type="text" className="form-control" id="last_name" name="last_name" required/>
-                        </div>
-
-                        <div className="form-group mb-3">
-                        <label className="mb-2"><strong>{t("PhoneNumber")}</strong></label>
-                            <input type="text" className="form-control" id="phone_number" name="phone_number" required pattern="[0-9]{3}-[0-9]{3}-[0-9]{4}" title='Please enter phone number in the following format: 000-000-0000.'/>
-                        </div>
-
-                        <div className="form-group mt-3">
-                            <button type="submit" class={styles.submitbutton}>{t("Complet")}</button>
-                        </div>
-                    </form>
+            <div className="card mt-5">
+              <form className="card-body" action="/api/forms/form" method="POST">
+                <input type="hidden" className="form-control" value={data?.email} id="email" name="email" />
+                <div className="form-group mb-3">
+                  <label className="mb-2"><strong>{t("FirstName")}</strong></label>
+                  <input type="text" className="form-control" defaultValue={data?.first_name} id="first_name" name="first_name" required />
                 </div>
+
+                <div className="form-group mb-3">
+                  <label className="mb-2"><strong>{t("LastName")}</strong></label>
+                  <input type="text" className="form-control" defaultValue={data?.last_name} id="last_name" name="last_name" required />
+                </div>
+
+                <div className="form-group mb-3">
+                  <label className="mb-2"><strong>{t("PhoneNumber")}</strong></label>
+                  <input type="text" className="form-control" defaultValue={data?.phone_number} id="phone_number" name="phone_number" required pattern="[0-9]{3}-[0-9]{3}-[0-9]{4}" title='Please enter phone number in the following format: 000-000-0000.' />
+                </div>
+
+                <div className="form-group mt-3">
+                  <button type="submit" class={styles.submitbutton}>{t("Update")}</button>
+                </div>
+              </form>
             </div>
-            <p class={styles.note}>*Note that completing the account creation will bring you back to the log in screen.</p>
+          </div>
         </main>
-  
+
         <footer>
           <p>Created By Carlo Staltari, Mohaned Bouzaidi & Yan Burton
-          <br />
-          Champlain College ECP Final Project 2022-2023</p>
+            <br />
+            Champlain College ECP Final Project 2022-2023</p>
         </footer>
 
         <style jsx>{`
@@ -211,7 +210,7 @@ export default function UserInfo({ user, data }) {
         }
       `}</style>
 
-      <style jsx global>{`
+        <style jsx global>{`
         html,
         body {
           padding: 0;
@@ -224,8 +223,8 @@ export default function UserInfo({ user, data }) {
           box-sizing: border-box;
         }
       `}
-      </style>
-    </div>)
+        </style>
+      </div>)
   )
 }
 
@@ -239,7 +238,9 @@ export const getServerSideProps = withPageAuthRequired({
       email: session.user.email
     });
     // access the user session
-    
+
+
+
     return { props: { data: JSON.parse(JSON.stringify(post)) } };
   }
 });
