+++ nextjs-blog/pages/project_submission.js
@@ -1,7 +1,8 @@
 import Head from 'next/head'
 import styles from '../styles/Home.module.css';
 import clientPromise from "../lib/mongodb";
-import Profile from './profile';
+// import Profile from './profile';
+import Profile from './/profile/index';
 import React, { useState } from "react";
 import { useTranslation } from "react-i18next";
 import "../src/Translation/i18n";
